bug修改

+++ mirco_service/store/src/main/java/com/cn/zbin/store/dto/ProductPriceExample.java
@@ -405,52 +405,52 @@ public class ProductPriceExample {
             return (Criteria) this;
         }
 
-        public Criteria andDaysEqualTo(Integer value) {
+        public Criteria andDaysEqualTo(Long value) {
             addCriterion("days =", value, "days");
             return (Criteria) this;
         }
 
-        public Criteria andDaysNotEqualTo(Integer value) {
+        public Criteria andDaysNotEqualTo(Long value) {
             addCriterion("days <>", value, "days");
             return (Criteria) this;
         }
 
-        public Criteria andDaysGreaterThan(Integer value) {
+        public Criteria andDaysGreaterThan(Long value) {
             addCriterion("days >", value, "days");
             return (Criteria) this;
         }
 
-        public Criteria andDaysGreaterThanOrEqualTo(Integer value) {
+        public Criteria andDaysGreaterThanOrEqualTo(Long value) {
             addCriterion("days >=", value, "days");
             return (Criteria) this;
         }
 
-        public Criteria andDaysLessThan(Integer value) {
+        public Criteria andDaysLessThan(Long value) {
             addCriterion("days <", value, "days");
             return (Criteria) this;
         }
 
-        public Criteria andDaysLessThanOrEqualTo(Integer value) {
+        public Criteria andDaysLessThanOrEqualTo(Long value) {
             addCriterion("days <=", value, "days");
             return (Criteria) this;
         }
 
-        public Criteria andDaysIn(List<Integer> values) {
+        public Criteria andDaysIn(List<Long> values) {
             addCriterion("days in", values, "days");
             return (Criteria) this;
         }
 
-        public Criteria andDaysNotIn(List<Integer> values) {
+        public Criteria andDaysNotIn(List<Long> values) {
             addCriterion("days not in", values, "days");
             return (Criteria) this;
         }
 
-        public Criteria andDaysBetween(Integer value1, Integer value2) {
+        public Criteria andDaysBetween(Long value1, Long value2) {
             addCriterion("days between", value1, value2, "days");
             return (Criteria) this;
         }
 
-        public Criteria andDaysNotBetween(Integer value1, Integer value2) {
+        public Criteria andDaysNotBetween(Long value1, Long value2) {
             addCriterion("days not between", value1, value2, "days");
             return (Criteria) this;
         }
